feat: added page for llm response
